投稿機能完成 & request_specの作成

+++ backend/app/controllers/api/v1/posts_controller.rb
@@ -9,19 +9,31 @@ class Api::V1::PostsController < ApplicationController
   end
 
   def create
-    post = Post.new(post_params)
-    if post.save
-      render json: post
+    if current_api_v1_user
+        post = Post.new(post_params)
+      if post.save
+        render json: post
+      else
+        render json: post.errors, status: 422
+      end
     else
-      render json: post.erros, status: 422
+      render json: {}, status: 422
     end
   end
 
   def destroy
-    post = Post.find(params[:id])
-    post.destroy
-    render json: post
+    post = Post.find_by(id: params[:id])
+    if post.present? && post.user_id == current_api_v1_user.id
+      post.destroy
+      render json: post, status: :ok
+    else
+      render_json_bad_request_with_custom_errors(
+        '投稿idが不正です',
+        '自分の投稿のみ削除する事ができます'
+      )
+    end
   end
+
 
   private
     def post_params
